ITFF_Project/requirements.txt
numpy>=1.26.0
pandas>=2.1.0
ta>=0.11.0
scikit-learn>=1.3.0
tensorflow>=2.14.0
torch>=2.1.0
matplotlib>=3.8.0
plotly>=5.18.0
dash>=2.14.0
requests>=2.31.0
flask>=3.0.0
fastapi>=0.103.0
uvicorn[standard]>=0.23.0
yahoo_fin>=0.8.9.1
python-dotenv>=1.0.0
pydantic>=2.3.0
transformers>=4.39.0
prometheus-client>=0.20.0
